+++ backend/models/user_usage.py
@@ -52,14 +52,14 @@ class UserUsage(UserIdentity):
 
         return request
 
-    def get_user_daily_usage(self, date):
+    def get_user_monthly_usage(self, date):
         """
         Fetch the user daily usage from the database
         """
         posthog = PostHogSettings()
-        request = self.supabase_db.get_user_requests_count_for_day(self.id, date)
+        request = self.supabase_db.get_user_requests_count_for_month(self.id, date)
         posthog.set_user_properties(
-            self.id, "DAILY_USAGE", {"daily_chat_usage": request}
+            self.id, "MONTHLY_USAGE", {"monthly_chat_usage": request}
         )
 
         return request
@@ -68,11 +68,12 @@ class UserUsage(UserIdentity):
         """
         Increment the user request count in the database
         """
-        current_requests_count = self.supabase_db.get_user_requests_count_for_day(
+        current_requests_count = self.supabase_db.get_user_requests_count_for_month(
             self.id, date
         )
-
+        logger.info(current_requests_count)
         if current_requests_count == 0:
+            logger.info("Request count is 0, creating new record")
             if self.email is None:
                 raise ValueError("User Email should be defined for daily usage table")
             self.supabase_db.create_user_daily_usage(
@@ -84,7 +85,6 @@ class UserUsage(UserIdentity):
         self.supabase_db.increment_user_request_count(
             user_id=self.id,
             date=date,
-            current_requests_count=current_requests_count,
             number=number,
         )